simple-tdd-demo/src/test/java/com/farhad/example/parse_query_string/QueryStringTest.java
package com.farhad.example.parse_query_string;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertThrows;

import org.junit.jupiter.api.Test;

public class QueryStringTest {

    @Test
    public void testOneNameValuePair() {
        QueryString query = new QueryString("name=value");
        assertEquals(1, query.count());
        assertEquals("value", query.valueFor("name"));
    }

    @Test
    public void testNoNameValuePairs() {
        QueryString query = new QueryString("");
        assertEquals(0, query.count());
    }

    @Test
    public void testNull() {
        assertThrows(NullPointerException.class, 
        () -> {
            new QueryString(null);
        });
    }
}
